requirements.txt
colorama
pymongo
pyfiglet
pynput
python-decouple